Fixes setuptools<=66.x to use gym==0.21.0 (#91)
Setuptools 67.0 onwards does not support malformed strings. As mentioned in their release notes, users must conform to PEP specifications strictly (i.e. will result in build errors if not complied). 

The MR fixes the `setuptools` package version to be <=66, so that the users can still continue using the gym==0.21.0. Though long-term we should switch to using Gymnasium. Additionally, it removes the malformed strings in Orbit packages to follow the new `setuptools`.