Fixes bug in deeper inheritance in configclass (#94)
# Description

This MR fixes the issue with multiple inheritances in the
`omni.isaac.orbit.utils.configclass` decorator. Earlier, if the
inheritance tree was more than one level deep and the lowest-level
configuration class was not updating its values from the middle-level
classes.

This was happening because in the middle-level classes, the default
values were being updated to dataclass fields and were not visible
anymore in the dictionary of the class itself. The configclass now
checks for their values in the `__dataclass_fields__` of the class as
well.

This was a regression in behavior but has been fixed now with a unit
test to capture deeper inheritance in configclass.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file