Adds terrain type directly to the terrain importer class (#95)
# Description

* Changed the behavior of the
`omni.isaac.orbit.terrains.TerrainImporter` class. It now expects the
terrain type to be specified in the configuration object. This allows
the user to specify everything in the configuration object and not have
to do an explicit call to import a terrain.
* Fixed setting of quaternion orientations inside the
`omni.isaac.orbit.markers.Visualizationmarkers` class. Earlier, the
orientation was being set into the point instancer in the wrong order
(`wxyz` instead of `xyzw`).

## Type of change

- Bug fix (non-breaking change which fixes an issue)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file