Fixes ground color and setting of ground transformation (#224)
# Description

This MR fixes the color of the ground in Orbit. It is now black since we
like that color more.

It also fixes the setting of the translation of the ground plane.
Earlier, it was taking z height from the configuration. Instead, now it
expects to get the full translation as an argument to its function. This
is needed to make it consistent with interactive scene designing API.

## Type of change

- Bug fix (non-breaking change which fixes an issue)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Screenshot


![new_ground](https://github.com/isaac-orbit/orbit/assets/12863862/8e729cb4-875f-476f-aa32-39cab38ad925)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file