Modifies the project license and adds CONTRIBUTORS.md (#169)
# Description

This MR adds the `CONTRIBUTORS.md` file, which lists the following three
categories:

* **Developers**: Individuals or organizations that are actively
maintaining and developing the project. They have "ownership" of the
project.
* **Contributors**: Individuals or organizations who are providing
contributions to the project but are not directly involved in its maintenance.
* **Acknowledgements**: Individuals or organizations who are involved in
discussions but have not contributed through code to the project.

All files henceforth use the copyright to the "ORBIT Project Developers"
instead of individuals or organizations.

When sending an MR, the user needs to add their name to this list. This
is mandatory for license agreements and is meant to protect
organizations.

## Type of change

- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file