Set livestream enabled to false unless it is otherwise specified (#374)
From Isaac `2023.1.0-hotfix.1` to `2023.1.1`, `/app/livestream/enabled`
started to default to True. This breaks our logic w/r/t
`SimulationContext.render()`, which is now treating a headless instance
as one that needs full rendering. This causes problems with multiple
SimulationContext calls, and ultimately this bottoms out with this
[line](https://github.com/isaac-orbit/orbit/blob/750c13add886b1decdf76d5e33a0f69ba792bda8/source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/simulation_context.py#L563)
sending us into an infinite rendering cycle from which i.e. unittest
cannot exit.

I fix this problem by setting `/app/livestream/enabled` to explicitly
False in AppLauncher unless otherwise requested. This corrects the hang.

Ultimately, the answer to this is #238, but in the meantime we need a
fix or this is a showstopper for unit testing.

Fixes #323

- Bug fix (non-breaking change which fixes an issue)

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have run all the tests with `./orbit.sh --test` and they pass
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there