Fixes mode-based checks inside the `EventManager.apply` call (#777)
# Description

Noticed that we were checking for function arguments inside a for-loop,
which isn't necessary. Moved this check outside to make it simpler to
read the code.

Also noticed a small corner case in the event manager when reset is
called and `env_ids` is None. In that case, it would bypass the check
for min steps between reset and directly apply the term to the
environment. I am not sure if that was intentional. if so, I can revert
the behavior.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there