Adds configurable modifiers to observation manager (#830)
# Description

The goal of this PR is to introduce expanded functionality to
observation corruption/realism. This is done by introducing the concept
of configurable modifiers. This creates an architecture that can apply a
list of custom modifiers to observation data giving flexibility to order
of operations. Existing modifiers for clip, scale, bias, and noise can
be used. This does so without changing the current functionality. This
PR also includes two new modifiers that are stateful (i.e. output is
dependent on previous state) that utilize a Callable class. They are a
`DigitalFilter` and an `Integrator`.

Summary of Additions:
- Adds `ModifierCfg` and `Modifier`
- Adds example modifiers functions for bias, scale, clip, noise
- Adds example modifiers classes: DigitalFilter, Integrator
- Adds functionality to `ObservationManager` to prepare and handle
modifiers of observation data
- Adds unit tests for modifiers in test_modifiers.py
- Adds test to test_observation_manager.py to test integration with
modifiers

## Type of change

- New feature (non-breaking change which adds functionality)
- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: jtigue-bdai <166445701+jtigue-bdai@users.noreply.github.com>
Co-authored-by: Mayank Mittal <mittalma@leggedrobotics.com>