Moves asset path from AppLauncher to app files (#228)
In a recent commit, Isaac Lab extensions were added to the app files as
dependencies in order to support installing Isaac Lab modules from pip.
However, this introduced a new issue where the assets module may be
loaded prior to the asset base path setting was set in AppLauncher,
causing assets to not be found. This causes the first runs of Isaac Lab
scripts to fail as the carbonite setting we look for in the assets
script has not been populated yet.

To fix this, this PR moves the definition of the asset path from
AppLauncher to the app files. This ensures that the carbonite setting is
set prior to any code being executed.

<!-- As you go through the list, delete the ones that are not
applicable. -->

- Bug fix (non-breaking change which fixes an issue)

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [ ] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->